third_party/WebKit/LayoutTests/imported/web-platform-tests/html/browsers/browsing-the-web/scroll-to-fragid/scroll-frag-percent-encoded-expected.txt

This is a testharness.js-based test.
FAIL Fragment Navigation: fragment id should be percent-decoded assert_equals: expected 400 but got 300
Harness: the test ran to completion.
